# gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/resources/openapi/openapi-apis.yaml
openapi: 3.0.3
info:
    title: Gravitee.io APIM - Management API - APIs
    description: |-
        This is the OpenAPI specification for our new version of APIM Management API.
    contact:
        email: team-apim@graviteesource.com
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    version: 2.0.0
security:
    - BasicAuth: []
    - CookieAuth: []
servers:
    - url: "{protocol}://{managementAPIHost}/management/v2"
      description: APIM Management API v2 - Default base URL
      variables:
          protocol:
              description: The protocol you want to use to communicate with the mAPI
              default: https
              enum:
                  - https
                  - http
          managementAPIHost:
              description: The domain of the server hosting your Management API
              default: localhost:8083
    - url: "{protocol}://{managementAPIHost}/management/v2/organizations/{orgId}"
      description: APIM Management API v2 - Base URL to target specific organizations
      variables:
          protocol:
              description: The protocol you want to use to communicate with the mAPI
              default: https
              enum:
                  - https
                  - http
          managementAPIHost:
              description: The domain of the server hosting your Management API
              default: localhost:8083
          orgId:
              description: The unique ID of your organization
              default: DEFAULT
...
paths:
...
    /environments/{envId}/integrations/{integrationId}/assets:
        parameters:
            - $ref: "#/components/parameters/envIdParam"
            - $ref: "#/components/parameters/integrationIdParam"
        get:
            tags:
                - Integration
            summary: Get a list of assets for a specific Integration
            description: |-
                Get a list of assets for a specific Integration.

                User must have at least one management (create, update, or delete) permission on the Integration to access it.
            operationId: getIntegrationAssets
            responses:
                "200":
                    $ref: "#/components/responses/IntegrationAssetsResponse"
                default:
                    $ref: "#/components/responses/Error"
    /environments/{envId}/integrations/{integrationId}/assets/_import:
        parameters:
            - $ref: "#/components/parameters/envIdParam"
            - $ref: "#/components/parameters/integrationIdParam"
        post:
            tags:
                - Integration
            summary: Import a list of assets for a specific Integration
            description: |-
                Import a list of assets for a specific Integration.

                User must have at least one management (create, update, or delete) permission on the Integration to access it.
            operationId: importIntegrationAssets
            requestBody:
                content:
                    application/json:
                        schema:
                            type: array
                            items:
                                $ref: "#/components/schemas/IntegrationAsset"
                required: true
            responses:
                "200":
                    $ref: "#/components/responses/IntegrationAssetsResponse"
                default:
                    $ref: "#/components/responses/Error"

    # API Subscriptions
...
components:
    schemas:
...
        Error:
            type: object
            properties:
                httpStatus:
                    type: integer
                    format: int32
                    description: The error code
                    example: 400
                message:
                    type: string
                    description: The error message
                    example: Bad request
                technicalCode:
                    type: string
                    description: A technical code to identify the error
                    example: invalid.import.definition
                parameters:
                    type: object
                    description: A map of parameters to be used in the error message
                    additionalProperties:
                        type: string
                details:
                    type: array
                    description: A list of details about the error
                    items:
                        type: object
                        properties:
                            message:
                                type: string
                                description: The error message
                                example: Bad request
                            location:
                                type: string
                                description: The json path of the field in error.
                                example: updateApi.properties[0].key
                            invalidValue:
                                type: object
                                description: The invalid value.
...
        IntegrationAsset:
            type: object
            properties:
                id:
                    type: string
                    description: Id of the asset
                name:
                    type: string
                    description: Name of the asset
                version:
                    type: string
                    description: Version of the asset
                description:
                    type: string
                    description: Description of the asset
                type:
                    type: string
                    description: Type of the asset
                    enum:
                        - OPENAPI
                        - ASYNCAPI
        Links:
            description: List of links for pagination
            properties:
                self:
                    type: string
                    description: Link to current resource
                first:
                    type: string
                    description: In a paginated response, link to the first page
                last:
                    type: string
                    description: In a paginated response, link to the last page
                previous:
                    type: string
                    description: In a paginated response, link to the previous page. Maybe null if current is the first page
                next:
                    type: string
                    description: In a paginated response, link to the next page. Maybe null if current is the last page
...
        Pagination:
            description: Generic object to handle pagination data.
            type: object
            properties:
                page:
                    type: integer
                    description: The current page.
                perPage:
                    type: integer
                    description: The number of items requested per page.
                pageCount:
                    type: integer
                    description: The total number of pages.
                pageItemsCount:
                    type: integer
                    description: The number of items for the current page.
                totalCount:
                    type: integer
                    format: int64
                    description: The total number of items.
...
    parameters:
...
        envIdParam:
            name: envId
            in: path
            required: true
            description: Id or Hrid (Human readable Id) of an environment.
            schema:
                type: string
                default: DEFAULT
...
        integrationIdParam:
            name: integrationId
            in: path
            required: true
            description: Id of an integration.
            schema:
                type: string
        ###############################
        # Pagination Query Parameters #
        ###############################
...
    responses:
...
        IntegrationAssetsResponse:
            description: Page of integration assets
            content:
                application/json:
                    schema:
                        title: "IntegrationAssetsResponse"
                        properties:
                            data:
                                description: List of assets for an integration.
                                type: array
                                items:
                                    $ref: "#/components/schemas/IntegrationAsset"
                            pagination:
                                $ref: "#/components/schemas/Pagination"
                            links:
                                $ref: "#/components/schemas/Links"
...
        Error:
            description: Generic error response
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"
...
    securitySchemes:
        BasicAuth:
            type: http
            scheme: basic
        CookieAuth:
            type: apiKey
            in: cookie
            name: Auth-Graviteeio-APIM